+++ src/react-reducer-provider.d.ts
@@ -4,14 +4,20 @@ import {
   ReactNode
 } from 'react'
 
+// Provider Component
+/////////////////////
+
+declare interface ProviderProps<STATE> {
+  name?: string | number;
+  initialState: STATE;
+  children: ReactNode;
+}
+
 // Reducer Component Definition
 ///////////////////////////////
 
-declare interface ReducerProps<STATE, REDUCER> {
-  name?: string | number;
+declare interface ReducerProps<STATE, REDUCER> extends ProviderProps<STATE> {
   reducer: REDUCER;
-  initialState: STATE;
-  children: ReactNode;
 }
 
 declare interface SyncReducer<STATE, ACTION> {
@@ -34,8 +40,35 @@ declare function AsyncReducerProvider<STATE, ACTION>(
   props: Readonly<AsyncReducerProps<STATE, ACTION>>
 ): ReactElement<AsyncReducerProps<STATE, ACTION>>
 
-// Reducer Consumption
-//////////////////////
+// Mapper Component Definition
+//////////////////////////////
+
+declare interface MapperProps<STATE, MAPPER> extends ProviderProps<STATE> {
+  mapper: MAPPER
+}
+
+declare interface SyncMapper<STATE, ACTION> {
+  (action: ACTION): STATE
+}
+
+declare interface SyncMapperProps<STATE, ACTION> extends MapperProps<STATE, SyncMapper<STATE, ACTION>> {}
+
+declare function SyncMapperProvider<STATE, ACTION>(
+  props: Readonly<SyncMapperProps<STATE, ACTION>>
+): ReactElement<SyncMapperProps<STATE, ACTION>>
+
+declare interface AsyncMapper<STATE, ACTION> {
+  (action: ACTION): Promise<STATE>
+}
+
+declare interface AsyncMapperProps<STATE, ACTION> extends MapperProps<STATE, AsyncMapper<STATE, ACTION>> {}
+
+declare function AsyncMapperProvider<STATE, ACTION>(
+  props: Readonly<AsyncMapperProps<STATE, ACTION>>
+): ReactElement<AsyncMapperProps<STATE, ACTION>>
+
+// Providers Consumption
+////////////////////////
 
 type Sync<T = any> = T
 type Async<T = any> = Promise<T>
@@ -44,18 +77,18 @@ declare interface Dispatcher<ACTION, DISPATCH extends Async | Sync = Sync<void>>
   (value: ACTION): DISPATCH
 }
 
-declare interface ReducerProviderValue<
+declare interface ProviderValue<
     STATE,
     ACTION,
-    DISPATCH extends Async<void | STATE> | Sync<void | STATE
-  > = Sync<void>> extends Array<any> {
+    DISPATCH extends Async<void | STATE> | Sync<void | STATE> = Sync<void>
+  > extends Array<any> {
   readonly 0: STATE;
   readonly 1: Dispatcher<ACTION, DISPATCH>;
 }
 
 declare function useReducer<STATE, ACTION, DISPATCH extends Async<void | STATE> | Sync<void | STATE> = Sync<void>>(
   name?: string | number
-): ReducerProviderValue<STATE, ACTION, DISPATCH>
+): ProviderValue<STATE, ACTION, DISPATCH>
 
 declare function useReducerState<STATE>(name?: string | number): STATE
 
@@ -72,21 +105,35 @@ declare interface Action<TYPE, DATA = undefined> {
 }
 
 export {
-  ReducerProps,
   SyncReducer,
   SyncReducerProps,
   SyncReducerProvider,
   AsyncReducer,
   AsyncReducerProps,
   AsyncReducerProvider,
+  SyncMapper,
+  SyncMapperProps,
+  SyncMapperProvider,
+  AsyncMapper,
+  AsyncMapperProps,
+  AsyncMapperProvider,
   Sync,
   Async,
   Dispatcher,
-  ReducerProviderValue,
+  ProviderValue,
   useReducer,
   useReducerState,
   useReducerDispatcher,
+  useReducer as useMapper,
+  useReducerState as useMapperState,
+  useReducerDispatcher as useMapperDispatcher,
   // Helpers
   //////////
-  Action
+  Action,
+  // Deprecated
+  /////////////
+  /**
+   * @deprecated since version 3.4.0, use 'ProviderValue' instead.
+   */
+  ProviderValue as ReducerProviderValue
 }