+++ src/useReducerState.js
@@ -1,8 +1,8 @@
 // Copyright (c) 2020 Gonzalo Müller Bravo.
 import * as React from 'react'
-import { captureReducerData } from './ReducerProvider'
+import { captureProvider } from './Providers'
 
 export function useReducerState(name) {
-  // Must `useContext` since it provides React triggering mechanism, using `captureReducerData(name).state` won't work.
-  return React.useContext(captureReducerData(name))[0]
+  // Must `useContext` since it provides React triggering mechanism, using `captureProvider(name).state` won't work.
+  return React.useContext(captureProvider(name))[0]
 }