+++ tests/typings/ts/typingTest.tsx
@@ -2,14 +2,19 @@
 import {
   Action,
   Async,
+  AsyncMapperProvider,
   AsyncReducerProvider,
+  Dispatcher,
+  ProviderValue,
+  Sync,
+  SyncMapperProvider,
+  SyncReducerProvider,
+  useMapper,
+  useMapperDispatcher,
+  useMapperState,
   useReducer,
   useReducerDispatcher,
-  useReducerState,
-  Dispatcher,
-  ReducerProviderValue,
-  Sync,
-  SyncReducerProvider
+  useReducerState
 } from '../../../src/react-reducer-provider'
 import React, {
   ReactElement,
@@ -111,7 +116,7 @@ function TestSingletonAsyncReducerProvider({ children }: {children: ReactNode}):
 }
 
 function TestSyncReducerMainHook(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string> = useReducer<TestState, string>('testNamedReducer')
+  const [ state, dispatch ]: ProviderValue<TestState, string> = useReducer<TestState, string>('testNamedReducer')
   return (
     <button onClick={(): void => dispatch('ACTION1')}>
       Child{state.lastAction}
@@ -120,7 +125,7 @@ function TestSyncReducerMainHook(): ReactElement {
 }
 
 function TestAsyncReducerMainHook(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string, Async> = useReducer<TestState, string, Async>('testNamedReducer')
+  const [ state, dispatch ]: ProviderValue<TestState, string, Async> = useReducer<TestState, string, Async>('testNamedReducer')
   const someFunc = () => {}
   return (
     <button onClick={async (): Promise<void> => await dispatch('ACTION1')
@@ -132,7 +137,7 @@ function TestAsyncReducerMainHook(): ReactElement {
 }
 
 function TestSyncReducerMainHookWithReturn(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string, Sync<TestState>> = useReducer<TestState, string, Sync<TestState>>('testNamedReducer')
+  const [ state, dispatch ]: ProviderValue<TestState, string, Sync<TestState>> = useReducer<TestState, string, Sync<TestState>>('testNamedReducer')
   return (
     <button onClick={(): TestState => dispatch('ACTION1')}>
       Child{state.lastAction}
@@ -141,7 +146,7 @@ function TestSyncReducerMainHookWithReturn(): ReactElement {
 }
 
 function TestAsyncReducerMainHookWithReturn(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string, Async<TestState>> = useReducer<TestState, string, Async<TestState>>('testNamedReducer')
+  const [ state, dispatch ]: ProviderValue<TestState, string, Async<TestState>> = useReducer<TestState, string, Async<TestState>>('testNamedReducer')
   const someFunc = (value: TestState) => {}
   return (
     <button onClick={async (): Promise<void> => await dispatch('ACTION1')
@@ -250,7 +255,7 @@ function TestNumberedAsyncReducerProvider({ children }: {children: ReactNode}): 
 }
 
 function TestNumberedSyncReducerMainHook(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string> = useReducer<TestState, string>(0)
+  const [ state, dispatch ]: ProviderValue<TestState, string> = useReducer<TestState, string>(0)
   return (
     <button onClick={(): void => dispatch('ACTION1')}>
       Child{state.lastAction}
@@ -259,7 +264,7 @@ function TestNumberedSyncReducerMainHook(): ReactElement {
 }
 
 function TestNumberedAsyncReducerMainHook(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string, Async> = useReducer<TestState, string, Async>(0)
+  const [ state, dispatch ]: ProviderValue<TestState, string, Async> = useReducer<TestState, string, Async>(0)
   const someFunc = () => {}
   return (
     <button onClick={async (): Promise<void> => await dispatch('ACTION1')
@@ -271,7 +276,7 @@ function TestNumberedAsyncReducerMainHook(): ReactElement {
 }
 
 function TestNumberedSyncReducerMainHookWithReturn(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string, Sync<TestState>> = useReducer<TestState, string, Sync<TestState>>(0)
+  const [ state, dispatch ]: ProviderValue<TestState, string, Sync<TestState>> = useReducer<TestState, string, Sync<TestState>>(0)
   return (
     <button onClick={(): TestState => dispatch('ACTION1')}>
       Child{state.lastAction}
@@ -280,7 +285,7 @@ function TestNumberedSyncReducerMainHookWithReturn(): ReactElement {
 }
 
 function TestNumberedAsyncReducerMainHookWithReturn(): ReactElement {
-  const [ state, dispatch ]: ReducerProviderValue<TestState, string, Async<TestState>> = useReducer<TestState, string, Async<TestState>>(0)
+  const [ state, dispatch ]: ProviderValue<TestState, string, Async<TestState>> = useReducer<TestState, string, Async<TestState>>(0)
   const someFunc = (value: TestState) => {}
   return (
     <button onClick={async (): Promise<void> => await dispatch('ACTION1')
@@ -334,6 +339,145 @@ function TestNumberedSyncReducerDispatcherHookWithReturn(): ReactElement {
 
 function TestNumberedAsyncReducerDispatcherHookWithReturn(): ReactElement {
   const theDispatcher: Dispatcher<number, Async<string>> = useReducerDispatcher<number, Async<string>>(0)
+  const someFunc = (value: string) => value.toLowerCase()
+  return (
+    <button onClick={(): Promise<string> => theDispatcher(123)
+      .then(someFunc)
+    }>
+      Children
+    </button>
+  )
+}
+
+function TestNumberedSyncMapperProvider({ children }: {children: ReactNode}): ReactElement {
+  function map(action: string): TestState {
+    switch (action) {
+      case 'ACTION1':
+        return {
+          lastAction: 1
+        }
+      default:
+        return initialState
+    }
+  }
+  return (
+    <SyncMapperProvider
+      name={0}
+      mapper={map}
+      initialState={initialState}
+    >
+      {children}
+    </SyncMapperProvider>
+  )
+}
+
+function TestNumberedAsyncMapperProvider({ children }: {children: ReactNode}): ReactElement {
+  async function map(action: string): Promise<TestState> {
+    switch (action) {
+      case 'ACTION1':
+        return {
+          lastAction: await Promise.resolve(1)
+        }
+      default:
+        return initialState
+    }
+  }
+  return (
+    <AsyncMapperProvider
+      name={0}
+      mapper={map}
+      initialState={initialState}
+    >
+      {children}
+    </AsyncMapperProvider>
+  )
+}
+
+function TestNumberedSyncMapperMainHook(): ReactElement {
+  const [ state, dispatch ]: ProviderValue<TestState, string> = useMapper<TestState, string>(0)
+  return (
+    <button onClick={(): void => dispatch('ACTION1')}>
+      Child{state.lastAction}
+    </button>
+  )
+}
+
+function TestNumberedAsyncMapperMainHook(): ReactElement {
+  const [ state, dispatch ]: ProviderValue<TestState, string, Async> = useMapper<TestState, string, Async>(0)
+  const someFunc = () => {}
+  return (
+    <button onClick={async (): Promise<void> => await dispatch('ACTION1')
+      .then(someFunc)
+    }>
+      Child{state.lastAction}
+    </button>
+  )
+}
+
+function TestNumberedSyncMapperMainHookWithReturn(): ReactElement {
+  const [ state, dispatch ]: ProviderValue<TestState, string, Sync<TestState>> = useMapper<TestState, string, Sync<TestState>>(0)
+  return (
+    <button onClick={(): TestState => dispatch('ACTION1')}>
+      Child{state.lastAction}
+    </button>
+  )
+}
+
+function TestNumberedAsyncMapperMainHookWithReturn(): ReactElement {
+  const [ state, dispatch ]: ProviderValue<TestState, string, Async<TestState>> = useMapper<TestState, string, Async<TestState>>(0)
+  const someFunc = (value: TestState) => {}
+  return (
+    <button onClick={async (): Promise<void> => await dispatch('ACTION1')
+      .then(someFunc)
+    }>
+      Child{state.lastAction}
+    </button>
+  )
+}
+
+function TestNumberedMapperStateHook(): ReactElement {
+  const theState: TestState = useMapperState<TestState>(0)
+  return (
+    <button>
+      Child{theState.lastAction}
+    </button>
+  )
+}
+
+function TestNumberedSyncMapperDispatcherHook(): ReactElement {
+  const theDispatcher: Dispatcher<string> = useMapperDispatcher<string>(0)
+  return (
+    <button onClick={(): void => theDispatcher('ACTION1')}>
+      Children
+    </button>
+  )
+}
+
+function TestNumberedAsyncMapperDispatcherHook(): ReactElement {
+  const theDispatcher: Dispatcher<string, Async> = useMapperDispatcher<string, Async>(0)
+  const someFunc = () => {}
+  return (
+    <button onClick={(): Promise<void> => theDispatcher('ACTION1')
+      .then(someFunc)
+    }>
+      Children
+    </button>
+  )
+}
+
+function TestNumberedSyncMapperDispatcherHookWithReturn(): ReactElement {
+  const theDispatcher: Dispatcher<number, Sync<string>> = useMapperDispatcher<number, Sync<string>>(0)
+  return (
+    <button onClick={(): void => {
+      const newState: string = theDispatcher(123)
+    }}>
+      Children
+    </button>
+  )
+}
+
+function TestNumberedAsyncMapperDispatcherHookWithReturn(): ReactElement {
+  const theDispatcher: Dispatcher<number, Async<string>> = useMapperDispatcher<number, Async<string>>(0)
   const someFunc = (value: string) => value.toLowerCase()
   return (
     <button onClick={(): Promise<string> => theDispatcher(123)